Move out search function in search.js
