+++ lib/index.js
@@ -1,2 +1,3 @@
 const app = require('./app');
-const styles = require('./styles.scss')
+const search = require('./search');
+const styles = require('./styles.scss');
